open-selector: Fix crash introduced in GDateTime port
commit acf14cc04f1e66c4c726d9a0b193bf39a28ca8b0 moved FileItems
over to using GDateTime instead of GTimeVal.

This reduced code complexity pretty nicely, but it unfortunately
introduced a bug, where the sort_items_by_mru function was treating
the FileItems themselves as GDateTime objects. That bug causes a
crash when opening files.

This commit fixes the problem by using the ->accessed members
instead of the file items themselves.

Closes: https://gitlab.gnome.org/GNOME/gedit/-/issues/465